Crate pjs_rs

Enums

ReturnValue

Js script return value, mapping types that can be deserialized as serde_json::Value or not, for the latest an string is returned witht the error message.

Functions

create_runtime_with_loader

Create a new runtime with the pjs extension built-in. Allow to pass a ModuleLoader to use, by default NoopModuleLoader is used.

run_file

Run a js/ts code from file in an isolated runtime with polkadotjs bundles embedded

run_file_with_wrap

Run a js/ts code from file in an isolated runtime with polkadotjs bundles embedded

run_js_code

run_ts_code