pjs_rs
Function run_file

pub async fn run_file(
    file_path: impl AsRef<Path>,
    json_args: Option<Vec<Value>>,
) -> Result<ReturnValue, AnyError>

Run a js/ts code from file in an isolated runtime with polkadotjs bundles embedded

The code runs without any wrapping clousure, and the arguments are binded to the pjs Object, available as pjs.arguments. If you want to to use top level await you will need to wrap your code like we do in run_file_with_wrap.

Example

Javascript file example

return pjs.api.consts.babe.epochDuration;

NOTE: To return a value you need to explicit call return <value>

Executing:

let resp = run_file("./testing/epoch_duration_rococo.js", None).await?;