pjs_rs
Function run_file_with_wrap

pub async fn run_file_with_wrap(
    file_path: impl AsRef<Path>,
    json_args: Option<Vec<Value>>,
) -> Result<ReturnValue, AnyError>

Run a js/ts code from file in an isolated runtime with polkadotjs bundles embedded

The code runs in a closure where the json_args are passed as arguments array and the polkadotjs modules (util, utilCrypto, keyring, types) are exposed from the global pjs Object. ApiPromise and WsProvider are also availables to easy access.

All code is wrapped within an async closure, allowing access to api methods, keyring, types, util, utilCrypto.

(async (arguments, ApiPromise, WsProvider, util, utilCrypto, keyring, types) => {
  ... any user code is executed here ...
})();

Example

Javascript file example

const api = await ApiPromise.create({ provider: new pjs.api.WsProvider('wss://rpc.polkadot.io') });
const parachains = (await api.query.paras.parachains()) || [];
console.log("parachain ids in polkadot:", parachains);
return parachains.toJSON();

NOTE: To return a value you need to explicit call return <value>

Executing:

let resp = run_file_with_wrap("./testing/query_parachains.js", None).await?;